refactor: move stuff around, simplify / code
