+++ src/web/questions/components/Stars.tsx
