+++ src/web/search/components/QueryForm.tsx
@@ -1,5 +1,3 @@
-import React from "react";
-
 interface Props {
   value: string;
   onChange: (v: string) => void;
